test/requirements.txt
python-dotenv
pyserial